% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_vboot.R
\name{AIC.glmnet}
\alias{AIC.glmnet}
\title{Calculate AIC with a glmnetfit object}
\usage{
\method{AIC}{glmnet}(fit)
}
\arguments{
\item{fit}{object fit with glmnet}
}
\value{

}
\description{
Calculate AIC for glmnet
}
