% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_vboot.R
\name{plot.bootVal}
\alias{plot.bootVal}
\title{Plot for repeat_cv results}
\usage{
\method{plot}{bootVal}(x, n = length(unique(unlist(x$varImportance))), ...)
}
\arguments{
\item{x}{A vboot object}

\item{n}{variables to be displayed}

\item{...}{further arguments passed to plot}
}
\value{
validation
}
\description{
Plots a grid of slices from the estimates of the repeat_cv function
}
