% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_vboot.R
\name{print.bootVal}
\alias{print.bootVal}
\title{Print function}
\usage{
\method{print}{bootVal}(x, ...)
}
\arguments{
\item{x}{A vboot object}

\item{...}{Further arguments of generic function}
}
\value{
validation
}
\description{
internal function to print vboot object
}
