% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_vboot.R
\name{score_binary}
\alias{score_binary}
\title{Score predicted probabilities}
\usage{
score_binary(y, p)
}
\arguments{
\item{y}{vector containing binary outcome}

\item{p}{vector containing predicted probabilities}
}
\value{
A `data.frame` containing binomial deviance, mean-squared error
   (brier score), mean absolute error, AUC, intercept and slope from a
   logistic calibration model (glm(y ~ qlogis(p), family="binomial"))
}
\description{
Score predicted probabilities against observed outcomes
}
