% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_vboot.R
\name{score_linear}
\alias{score_linear}
\title{Score predictions from linear model}
\usage{
score_linear(y, p)
}
\arguments{
\item{y}{vector containing outcome}

\item{p}{vector containing predictions}
}
\value{
A `data.frame` containing R2, MSE, MAE, intercept and slope from a
   linear calibration model (lm(y ~ p))
}
\description{
Score predictions from linear model
}
