% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_vboot.R
\name{vboot.glmnet}
\alias{vboot.glmnet}
\title{Internal bootstraping validation logistic glmnet model}
\usage{
\method{vboot}{glmnet}(
  fit,
  x,
  y,
  s,
  gamma = NULL,
  nfolds = 5,
  B = 200,
  cv_replicates = 100,
  n_cores = max(1, parallel::detectCores() - 1),
  AIC_select = FALSE,
  save_ps = FALSE
)
}
\arguments{
\item{fit}{Object from glmnet fit}

\item{x}{A matrix of the predictors, each row is an observation vector.}

\item{y}{A vector of response variable. Should be a factor with two levels}

\item{s}{Value of the penalty parameter "lambda" selected from the original 'cv.glmnet'}

\item{gamma}{Value of "gamma" parameter selected for relaxed model}

\item{nfolds}{Number of folds for cross validation as in cv.glmnet}

\item{B}{Number of bootsrap samples}

\item{cv_replicates}{Number of replicates for the cross-validation step in 'cv.glmnet'}

\item{n_cores}{number of cores to use in parallel. Default detectCores()-1}

\item{AIC_select}{use AIC (instead of cross vaidation) to select lambda? This should have been done to select the original lambdq (default = FALSE)}

\item{save_ps}{save the predicted probabilities from the bootstrapped models? This is used for non-linear calibration curves (default = FALSE)}
}
\description{
Validate glmnet logistic regression using bootstrap.
}
