% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_vboot.R
\name{vboot.lm}
\alias{vboot.lm}
\title{Internal bootstraping validation linear model}
\usage{
\method{vboot}{lm}(
  fit,
  x = NULL,
  y = NULL,
  s = NULL,
  gamma = NULL,
  nfolds = NULL,
  B = 200,
  cv_replicates = NULL,
  n_cores = max(1, parallel::detectCores() - 1)
)
}
\arguments{
\item{fit}{Object from lm fit}

\item{x}{A matrix of the predictors, each row is an observation vector.}

\item{y}{A vector of response variable. It should be quantitative for lineal regression, a factor with two levels for logistic regression or a two-column matrix with columns named 'time' and 'status' for cox regression.}

\item{s}{Value of the penalty parameter "lambda" selected from the original 'cv.glmnet'.}

\item{gamma}{Value of "gamma" parameter selected for relaxed model}

\item{nfolds}{Number of folds for cross validation as in 'cv.glmnet'.}

\item{B}{Number of bootsrap samples}

\item{cv_replicates}{Number of replicates for the cross-validation step}

\item{n_cores}{number of cores to use in parallel. Default detectCores()-1}
}
\description{
Validate linear regression using bootstrap.
}
